pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ai-content-compliance-openenv"
version = "1.0.0"
description = "OpenEnv content moderation RL environment"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0.0",
    "openai>=1.12.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "websockets>=12.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]
